+++ src/lib/app-provider/hooks/useInitia.ts
@@ -1,0 +1,8 @@
+import { useCelatoneApp } from "../contexts";
+
+export const useInitia = () => {
+  const {
+    chainConfig: { chain },
+  } = useCelatoneApp();
+  return chain === "initia";
+};